mqttSubcriber/requirements.txt
certifi==2025.1.31
dnspython==2.7.0
motor==3.7.0
paho-mqtt==2.1.0
pymongo==4.11.3
python-dotenv==1.0.1
